requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-jobspy>=1.1.50
requests>=2.28.0
beautifulsoup4>=4.12.0
tls-client>=0.2.0
httpx>=0.24.0
tenacity>=8.2.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
